- removed unused p2p components
