+++ p2p/libp2p/netMessenger.go
@@ -266,7 +266,6 @@ func (netMes *networkMessenger) createPubSub(withMessageSigning bool) error {
 func (netMes *networkMessenger) createSharder(p2pConfig config.P2PConfig) error {
 	args := factory.ArgsSharderFactory{
 		PeerShardResolver:       &unknownPeerShardResolver{},
-		PrioBits:                p2pConfig.Sharding.PrioBits,
 		Pid:                     netMes.p2pHost.ID(),
 		MaxConnectionCount:      p2pConfig.Sharding.TargetPeerCount,
 		MaxIntraShardValidators: int(p2pConfig.Sharding.MaxIntraShardValidators),
