+++ p2p/libp2p/networksharding/factory/sharderFactory.go
@@ -14,7 +14,6 @@ var log = logger.GetOrCreate("p2p/networksharding/factory")
 // ArgsSharderFactory represents the argument for the sharder factory
 type ArgsSharderFactory struct {
 	PeerShardResolver       p2p.PeerShardResolver
-	PrioBits                uint32
 	Pid                     peer.ID
 	MaxConnectionCount      int
 	MaxIntraShardValidators int
@@ -27,14 +26,6 @@ type ArgsSharderFactory struct {
 // NewSharder creates new Sharder instances
 func NewSharder(arg ArgsSharderFactory) (p2p.CommonSharder, error) {
 	switch arg.Type {
-	case p2p.PrioBitsSharder:
-		log.Debug("using prio bits sharder",
-			"PrioBits", arg.PrioBits,
-		)
-		return networksharding.NewPrioBitsSharder(arg.PrioBits, arg.PeerShardResolver)
-	case p2p.SimplePrioBitsSharder:
-		log.Debug("using simple prio bits sharder")
-		return &networksharding.SimplePrioBitsSharder{}, nil
 	case p2p.ListsSharder:
 		log.Debug("using lists sharder",
 			"MaxConnectionCount", arg.MaxConnectionCount,
